Feat: Remove unnecessary additions and removals using TVM updates
This removes the now unnecessay remove and re add of tasks and constraints from the solver every time they are modified.
This is handled automatically by TVM using the changes made by https://github.com/jrl-umi3218/tvm/pull/53 and does not need to be tracked at the solver level anymore.